+++ src/main/java/com/mycom/myapp/naviya/global/controller/PageController.java
@@ -6,23 +6,23 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class PageController {
 
-    @GetMapping("/login")
-    public String loginP() {
+//    @GetMapping("/login")
+//    public String loginP() {
+//
+//        return "login";
+//    }
+//
+//    @GetMapping("/signup")
+//    public String signupP() {
+//
+//        return "signup";
+//    }
 
-        return "login";
-    }
-
-    @GetMapping("/signup")
-    public String signupP() {
-
-        return "signup";
-    }
-
-    @GetMapping("/usermypage")
-    public String userP() {
-
-        return "userpage";
-    }
+//    @GetMapping("/usermypage")
+//    public String userP() {
+//
+//        return "userpage";
+//    }
