+++ src/main/resources/static/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html xmlns:th="http://www.thymeleaf.org" lang="ko">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -99,7 +99,7 @@
         <div class="collapse navbar-collapse justify-content-end">
             <ul class="navbar-nav">
                 <li class="nav-item">
-                    <a class="nav-link" href="#">메인페이지</a>
+                    <a class="nav-link" href="/">메인페이지</a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" href="#">로그인</a>
