+++ src/main/resources/templates/main.html
@@ -6,7 +6,7 @@
 <body>
 <h1>MAIN페이지입니당</h1>
 <hr>
-<h1>Welcome, <span th:text="${id}">User</span>!</h1>
+<h1>Welcome, <span th:text="${email}">User</span>!</h1>
 
 <p>EMAIL: <span th:text="${email}">email</span></p>
 <p>Role: <span th:text="${role}">Role</span></p>
